scans/requirements.txt
django
django_rest_framework
GitPython
keras
requests
tensorflow